use size_t instead of int for a few situations Spacer now appears as rounded rect rather than dots

+++ src/ofxCvGui/Panels/Groups/Grid.cpp
@@ -61,8 +61,8 @@ namespace ofxCvGui {
 				bounds.height = panelHeight;
 
 				float y = 0;
-				int row = 0;
-				int col = 0;
+				size_t row = 0;
+				size_t col = 0;
 				if (this->heights.size() > 0) {
 					bounds.height = heights[0];
 				}
